#4 - wifi 서비스 구성

+++ src/main/java/com/zerobase/api/dto/WifiApiResponse.java
@@ -8,12 +8,12 @@ import java.util.List;
 @Getter
 public class WifiApiResponse {
 
-    private final Long totalCount;
+    private final int totalCount;
     private final WifiApiResultResponse result;
     private final List<WifiApiIntoDetailResponse> details;
 
     @Builder
-    private WifiApiResponse(Long totalCount, WifiApiResultResponse result, List<WifiApiIntoDetailResponse> details) {
+    private WifiApiResponse(int totalCount, WifiApiResultResponse result, List<WifiApiIntoDetailResponse> details) {
         this.totalCount = totalCount;
         this.result = result;
         this.details = details;
